cr100

+++ dsmg/public/packages/script/crossing.js
@@ -163,8 +163,10 @@ function overdueStation(){
 $("#data-list").on("click", ".del", function(){
 	$deleting = $(this);
 	var id = $deleting.closest("tr").find(".id").text();
-	$deleting.html('&nbsp;&nbsp;<i class="fa fa-spinner fa-spin fa-lg"></i>&nbsp;&nbsp;');
-	delete_data(id);
+	alertify.confirm('Please confirm your delete!').set('onok', function(closeEvent){
+				$deleting.html('&nbsp;&nbsp;<i class="fa fa-spinner fa-spin fa-lg"></i>&nbsp;&nbsp;');
+				delete_data(id);
+	});
 });
 function delete_data(id){
 	$.ajax({
